@@ -23,12 +23,12 @@ nox.options.reuse_existing_virtualenvs = True
 
 @nox.session(
     python=[
-        "3.7",
         "3.8",
         "3.9",
         "3.10",
         "3.11",
         "3.12",
+        "3.13",
         "pypy3.8",
         "pypy3.9",
         "pypy3.10",
